同步钉钉、userInfo数据分布式锁

+++ source/urc-bp-impl/src/main/java/com/yks/urc/bp/impl/UserBp.java
@@ -5,6 +5,7 @@ import com.yks.urc.entity.UserInfo;
 import com.yks.urc.entity.UrcUserDo;
 import com.yks.urc.fw.HttpUtility;
 import com.yks.urc.fw.StringUtility;
+import com.yks.urc.lock.DistributedReentrantLock;
 import com.yks.urc.mapper.IUrcUserMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,6 +39,8 @@ public class UserBp {
 
     @Autowired
     private IUrcUserMapper userMapper;
+    
+	DistributedReentrantLock lock = new DistributedReentrantLock("SynUserFromUserInfo");
 
     /**
      * 同步UserInfo数据
@@ -47,36 +50,44 @@ public class UserBp {
      */
     @Transactional(rollbackFor = Exception.class)
     public void SynUserFromUserInfo(String username) {
-        List<UserInfo> userInfoList = this.getUserInfo();
-        UrcUserDo userDo = new UrcUserDo();
-        for (UserInfo user : userInfoList) {
-            userDo.setUsername(user.username);
-            userDo.setDingUserId(user.ding_userid);
-            userDo.setCreateBy(username);
-            userDo.setModifiedBy(username);
-            try {
-                userDo.setActiveTime(StringUtility.stringToDate(user.date_joined,"yyyy-MM-dd HH:mm:ss"));
-                //0 表示启用,1表示禁用
-                if ("66050".equals(user.userAccountControl)) {
-                    userDo.setIsActive(1);
-                } else if ("66048".equals(user.userAccountControl) || "512".equals(user.userAccountControl)) {
-                    userDo.setIsActive(0);
-                } else {
-                    userDo.setIsActive(Integer.parseInt(user.userAccountControl));
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-        //传入手动同步的创建人员
-        userDo.setCreateBy(username);
-        userDo.setModifiedBy(username);
-        List<UrcUserDo> userDoList = new ArrayList<>();
-        userDoList.add(userDo);
-        //先清理数据表
-        userMapper.deleteUrcUser();
-        logger.info("清理完成,开始同步");
-        userMapper.insertBatchUser(userDoList);
+    	if(lock.tryLock()){
+    		 List<UserInfo> userInfoList = this.getUserInfo();
+    	        UrcUserDo userDo = new UrcUserDo();
+    	        for (UserInfo user : userInfoList) {
+    	            userDo.setUsername(user.username);
+    	            userDo.setDingUserId(user.ding_userid);
+    	            userDo.setCreateBy(username);
+    	            userDo.setModifiedBy(username);
+    	            try {
+    	                userDo.setActiveTime(StringUtility.stringToDate(user.date_joined,"yyyy-MM-dd HH:mm:ss"));
+    	                //0 表示启用,1表示禁用
+    	                if ("66050".equals(user.userAccountControl)) {
+    	                    userDo.setIsActive(1);
+    	                } else if ("66048".equals(user.userAccountControl) || "512".equals(user.userAccountControl)) {
+    	                    userDo.setIsActive(0);
+    	                } else {
+    	                    userDo.setIsActive(Integer.parseInt(user.userAccountControl));
+    	                }
+    	    	        //传入手动同步的创建人员
+    	    	        userDo.setCreateBy(username);
+    	    	        userDo.setModifiedBy(username);
+    	    	        List<UrcUserDo> userDoList = new ArrayList<>();
+    	    	        userDoList.add(userDo);
+    	    	        //先清理数据表
+    	    	        userMapper.deleteUrcUser();
+    	    	        logger.info("清理完成,开始同步");
+    	    	        userMapper.insertBatchUser(userDoList);
+    	            } catch (Exception e) {
+    	                e.printStackTrace();
+    	            }finally{
+    					lock.unlock();
+    				}
+    	        }
+    	}else{
+	        logger.info("同步userInfo数据正在执行...,");
+		}
+    	
+       
     }
 
     /**
